+++ src/dto/UpdateUser.dto.ts
@@ -1,0 +1,6 @@
+interface UpdateUserDto {
+  nickname: string;
+  seasonId: number;
+}
+
+export default UpdateUserDto;
