+++ src/er-api/er-api-queue.service.ts
@@ -1,6 +1,7 @@
 import { InjectQueue } from '@nestjs/bull';
 import { Injectable } from '@nestjs/common';
 import { Queue } from 'bull';
+import UpdateUserDto from 'src/dto/UpdateUser.dto';
 import { User } from 'src/entities/User.entity';
 
 @Injectable()
@@ -20,12 +21,11 @@ export class ErApiQueueService {
     return jobResult;
   }
 
-  async getInfo(nickname: string, seasonId: number) {
-    const job = await this.erApiQueue.add(
-      'getInfo',
-      { nickname, seasonId },
-      { removeOnComplete: true, removeOnFail: true },
-    );
+  async updateUser(updateUserDto: UpdateUserDto) {
+    const job = await this.erApiQueue.add('updateUser', updateUserDto, {
+      removeOnComplete: true,
+      removeOnFail: true,
+    });
 
     const jobResult: User = await job.finished();
