+++ src/er-api/er-api.service.ts
@@ -1,52 +1,26 @@
 import { Injectable } from '@nestjs/common';
-import { InjectRepository } from '@nestjs/typeorm';
 import * as dayjs from 'dayjs';
-import { Game } from 'src/entities/Game.entity';
-import { GameUser } from 'src/entities/GameUser.entity';
-import { SeasonUser } from 'src/entities/SeasonUser.entity';
-import { User } from 'src/entities/User.entity';
-import {
-  BattleUserResult,
-  ErGameList,
-  ErSeasonUser,
-  ErUserInfo,
-} from 'src/types/erApiResponse';
-import { MatchingTeamMode } from 'src/types/matchingTeamMode';
-import erApiService from 'src/utils/erApiService';
-import { Repository } from 'typeorm';
+import delay from 'src/utils/delay';
+import { ErApiUserService } from './er-api-user.service';
+import { ErApiGameService } from './er-api-game.service';
+import UpdateUserDto from 'src/dto/UpdateUser.dto';
 dayjs().format();
 
 @Injectable()
 export class ErApiService {
   constructor(
-    @InjectRepository(User)
-    private readonly userRepository: Repository<User>,
-    @InjectRepository(Game)
-    private readonly gameRepository: Repository<Game>,
-    @InjectRepository(GameUser)
-    private readonly gameUserRepository: Repository<GameUser>,
-    @InjectRepository(SeasonUser)
-    private readonly seasonUserRepository: Repository<SeasonUser>,
+    private readonly userService: ErApiUserService,
+    private readonly gameService: ErApiGameService,
   ) {}
 
-  delay(ms: number): Promise<void> {
-    return new Promise((resolve) => {
-      setTimeout(resolve, ms);
-    });
-  }
-
-  async getUserInfo(nickname: string, seasonId: number): Promise<User> {
-    const user = await this.registerUser(nickname);
+  async updateUser(updateUserDto: UpdateUserDto) {
+    const { nickname, seasonId } = updateUserDto;
+    const user = await this.userService.registerUser(nickname);
     const now = dayjs();
 
-    await this.delay(1000);
+    await delay(1000);
 
     let next: string;
-
-    const userData = await this.userRepository.findOne({
-      where: { id: user.id },
-      relations: ['gameUsers', 'gameUsers.game'],
-    });
 
     let requestType: 'full' | 'fast' = 'full';
 
@@ -54,11 +28,11 @@ export class ErApiService {
       requestType = 'fast';
     }
 
-    const alreadyGameList = userData.gameUsers.map((gameUser) => {
+    const alreadyGameList = user.gameUsers.map((gameUser) => {
       return gameUser.game.id;
     });
 
-    let gameList = await this.registerGame(
+    let gameList = await this.gameService.registerGame(
       user,
       alreadyGameList,
       requestType,
@@ -67,8 +41,8 @@ export class ErApiService {
 
     while (gameList) {
       next = gameList;
-      await this.delay(1500);
-      gameList = await this.registerGame(
+      await delay(1500);
+      gameList = await this.gameService.registerGame(
         user,
         alreadyGameList,
         requestType,
@@ -76,169 +50,16 @@ export class ErApiService {
       );
     }
 
-    await this.delay(1000);
+    await delay(1000);
 
-    this.registerSeasonData(user, seasonId);
+    await this.userService.registerSeasonData(user, seasonId);
 
-    await this.delay(1000);
+    await delay(1000);
 
-    await this.userRepository.update(user.id, { lastUpdate: new Date() });
+    await this.userService.updatedTime(user.id);
 
-    return await this.userRepository.findOne({
-      where: { id: user.id },
-      relations: ['gameUsers', 'gameUsers.game'],
-    });
-  }
-
-  async registerUser(nickname: string, id?: number): Promise<User> {
-    const user = await this.userRepository.findOne({
-      where: { nickname },
-    });
-
-    if (user) {
-      return user;
-    }
-
-    if (id) {
-      const newUser = this.userRepository.create({
-        id,
-        nickname,
-      });
-
-      await this.userRepository.insert(newUser);
-
-      return newUser;
-    }
-
-    const userData = await erApiService.get<ErUserInfo>(
-      `/v1/user/nickname?query=${nickname}`,
-    );
-
-    const newUser = this.userRepository.create({
-      id: userData.data.user.userNum,
-      nickname: userData.data.user.nickname,
-      lastUpdate: new Date(),
-    });
-
-    await this.userRepository.insert(newUser);
-
-    return newUser;
-  }
-
-  async registerGame(
-    user: User,
-    alreadyGames: number[],
-    type: 'fast' | 'full',
-    next?: string,
-  ): Promise<string> {
-    const gameData = await erApiService.get<ErGameList>(
-      `v1/user/games/${user.id}${next ? `?next=${next}` : ''}`,
-    );
-
-    const gameList: BattleUserResult[] = [];
-
-    gameData.data.userGames.forEach((game) => {
-      if (!alreadyGames.includes(game.gameId)) {
-        gameList.push(game);
-      }
-    });
-
-    for (const game of gameList) {
-      const isGameExist = await this.gameRepository.findOne({
-        where: { id: game.gameId },
-      });
-
-      if (!isGameExist) {
-        const insertGameData = {
-          id: game.gameId,
-          startDtm: new Date(game.startDtm),
-          playTime: game.playTime,
-        };
-
-        const newGame = this.gameRepository.create(insertGameData);
-        await this.gameRepository.insert(newGame);
-      }
-
-      const existingGame = await this.gameRepository.findOne({
-        where: { id: game.gameId },
-      });
-
-      const insertGameUserData = {
-        KillCount: game.playerKill,
-        AnimalKillCount: game.monsterKill,
-        AssistCount: game.playerAssistant,
-        characterNum: game.characterNum,
-        DeathCount: game.playerDeaths,
-        TeamKillCount: game.teamKill,
-        user: user,
-        game: existingGame,
-        seasonId: game.seasonId,
-        id: `${game.userNum}-${game.gameId}-${game.seasonId}`,
-      };
-
-      const newGameUser = this.gameUserRepository.create(insertGameUserData);
-
-      await this.gameUserRepository.insert(newGameUser);
-    }
-
-    if (type === 'fast' && gameList.length < 10) {
-      return undefined;
-    }
-
-    return gameData.data.next;
-  }
-
-  async registerSeasonData(user: User, seasonId: number) {
-    const res = await erApiService.get<ErSeasonUser>(
-      `/v1/user/stats/${user.id}/${seasonId}`,
-    );
-
-    const seasonData = res.data.userStats;
-
-    if (!seasonData) {
-      return;
-    }
-
-    for (const season of seasonData) {
-      const id = `${season.userNum}-${season.seasonId}-${
-        MatchingTeamMode[season.matchingTeamMode]
-      }`;
-
-      const seasonUser = {
-        id,
-        averageAssists: season.averageAssistants,
-        averageHunts: season.averageHunts,
-        averageKills: season.averageKills,
-        averageRank: season.averageRank,
-        matchingTeamMode: season.matchingTeamMode,
-        mmr: season.mmr,
-        rank: season.rank,
-        rankSize: season.rankSize,
-        seasonId: season.seasonId,
-        top2Rate: season.top2 * 100,
-        top3Rate: season.top3 * 100,
-        totalGames: season.totalGames,
-        totalWins: season.totalWins,
-        user: user,
-      } as SeasonUser;
-
-      if (seasonUser.totalGames === 0) {
-        seasonUser.winRate = 0;
-      } else {
-        const winRate = (seasonUser.totalWins / seasonUser.totalGames) * 100;
-        seasonUser.winRate = Number(winRate.toFixed(2));
-      }
-
-      const isSeasonUserExist = await this.seasonUserRepository.findOne({
-        where: { id },
-      });
-
-      if (!isSeasonUserExist) {
-        const newSeasonUser = this.seasonUserRepository.create(seasonUser);
-        await this.seasonUserRepository.insert(newSeasonUser);
-      } else {
-        await this.seasonUserRepository.update(id, seasonUser);
-      }
-    }
+    return {
+      message: `User ${nickname} updated!`,
+    };
   }
 }
