+++ src/utils/delay.ts
@@ -1,0 +1,7 @@
+const delay = (ms: number): Promise<void> => {
+  return new Promise((resolve) => {
+    setTimeout(resolve, ms);
+  });
+};
+
+export default delay;
